build: run game tests on all main commits
super annoying, but in a monorepo if you selectively run your
coverage only on the commits which change something in your
subproject, that's fine and all... but when you then run
coverage on a PR, it only knows the commit from where that
branch forked, and if there isnt any coverage on that commit
(i.e. it wasn't a change that impacted your subproject), then
it has no idea what the basis is for coverage changes.

it's just not really a problem when you have 1 artifact per
repo.